scripts/requirements.txt
pymongo==3.11.3
requests==2.25.1
python-dotenv==0.20.0
